购买

+++ SpringMvcStudy/src/com/study/bean/RealUser.java
@@ -7,7 +7,7 @@ public class RealUser implements Serializable{
 	private String user_name;//用户名
 	private String user_code;//资金账号
 	private String user_password;//用户密码
-	private int user_asset;//用户资产
+	private double user_asset;//用户资产
 	private String reg_time;//注册时间
 
 
@@ -19,11 +19,11 @@ public class RealUser implements Serializable{
 		this.reg_time = reg_time;
 	}
 
-	public int getUser_asset() {
+	public double getUser_asset() {
 		return user_asset;
 	}
 
-	public void setUser_asset(int user_asset) {
+	public void setUser_asset(double user_asset) {
 		this.user_asset = user_asset;
 	}
